Move ai structures into companyai.h
